Moved veraPDF helpers into the extensions directory.
